Add Trailing .0 To Licence Declaration.
The SPDX licences must have changed slightly from the inception of this
package. The new Apache licences require a .0 for both v1 and v2. This
bumps this in the package.json.

Some build systems (Facebook's Rome) will hiccup/complain that these
licenses are not inline with those listed on the spdx.org listing.

SPDX listing:
https://spdx.org/licenses/